Componentize the Delete button on the person form
